requirements.txt
Flask==1.1.1
Werkzeug==0.16.0
Flask-cors==3.0.8
GoogleAppEngineCloudStorageClient==1.9.22.1
google-api-core==1.14.3
pyasn1==0.4.8
google-cloud-storage==1.23.0
pytest==5.3.1
google-cloud-logger==0.2.1
flask-google-cloud-logger==0.2.1
pika==1.1.0
amqp==2.5.2
Twisted==19.10.0
google-cloud-pubsub==1.1.0
google-cloud-core==1.2.0
